refactor: Remove blanks
