+++ src/main/java/com/cherryhouse/server/_core/config/AwsConfig.java
@@ -1,7 +1,10 @@
 package com.cherryhouse.server._core.config;
 
+import com.amazonaws.auth.AWSCredentials;
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
+import com.amazonaws.services.s3.AmazonS3;
+import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 import com.amazonaws.services.simpleemail.AmazonSimpleEmailService;
 import com.amazonaws.services.simpleemail.AmazonSimpleEmailServiceClientBuilder;
 import org.springframework.beans.factory.annotation.Value;
@@ -11,6 +14,9 @@ import org.springframework.context.annotation.Configuration;
 @Configuration
 public class AwsConfig {
 
+    //AWS에 접근 할 때, 어떤 IAM을 통해 접근하고, 어떤 지역에서 접근하는지를 설정하는 Configuration
+    //얻은 자격증명 객체를 이용해서 AmazonS3ClientBuilder를 통해 S3 Client를 가져옴
+
     @Value("${cloud.aws.credentials.access-key}")
     private String accessKey;
 
@@ -19,6 +25,15 @@ public class AwsConfig {
 
     @Value("${cloud.aws.region.static}")
     private String region;
+
+    @Bean
+    public AmazonS3 s3Builder(){
+        AWSCredentials basicAWSCredentials = new BasicAWSCredentials(accessKey,secretKey);
+
+        return AmazonS3ClientBuilder.standard()
+                .withCredentials(new AWSStaticCredentialsProvider(basicAWSCredentials))
+                .withRegion(region).build();
+    }
 
     @Bean
     public AmazonSimpleEmailService amazonSimpleEmailService(){
