+++ src/main/java/com/cherryhouse/server/user/UserRepository.java
@@ -1,11 +1,9 @@
 package com.cherryhouse.server.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
-@Repository
 public interface UserRepository extends JpaRepository<User, Long> {
 
     boolean existsByEmail(String email);
