+++ src/main/java/com/cherryhouse/server/user/UserService.java
@@ -9,20 +9,17 @@ import com.cherryhouse.server.style.Style;
 import com.cherryhouse.server.style.StyleRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
-import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class UserService {
-    private static final Logger log = LoggerFactory.getLogger(UserService.class);
+
     public final UserRepository userRepository;
     public final StyleRepository styleRepository;
     public final S3Service s3Service;
@@ -38,6 +35,30 @@ public class UserService {
 
         userRepository.save(user);
         return user;
+    }
+
+    @Transactional
+    public void uploadStyle(List<MultipartFile> file, String email) {
+        User user = findByEmail(email);
+        List<String> fileNames = s3Service.upload(file,"style");
+
+        fileNames.forEach( filePath ->{
+            Style style = Style.builder()
+                    .user(user)
+                    .imgUrl(filePath)
+                    .build();
+            styleRepository.save(style);
+        });
+    }
+
+    @Transactional
+    public void deleteStyle(String filePath, String email) {
+        findByEmail(email);
+        Long id = styleRepository.findByUserEmailAndFileName(email,filePath).orElseThrow(
+                ( )->new ApiException(ExceptionCode.INVALID_REQUEST_DATA,"해당 스타일이 존재하지 않습니다.")
+        );
+        s3Service.delete(filePath);
+        styleRepository.deleteById(id);
     }
 
     public void existsByEmail(String email){
@@ -57,30 +78,5 @@ public class UserService {
         return userRepository.findByEmail(email).orElseThrow(
                 ()-> new ApiException(ExceptionCode.USER_NOT_FOUND)
         );
-    }
-
-    @Transactional
-    public void uploadStyle(List<MultipartFile> file, String email) {
-        User user = findByEmail(email);
-        List<String> fileNames = s3Service.upload(file,"style");
-
-        fileNames.forEach( filePath ->{
-            Style style = Style.builder()
-                    .user(user)
-                    .imgUrl(filePath)
-                    .build();
-            styleRepository.save(style);
-        });
-    }
-
-    @Transactional
-    public void deleteStyle(String filePath, String email) {
-        User user = findByEmail(email);
-        Long id = styleRepository.findByUserEmailAndFileName(email,filePath).orElseThrow(
-                ( )->new ApiException(ExceptionCode.INVALID_REQUEST_DATA,"해당 스타일이 존재하지 않습니다.")
-        );
-        s3Service.delete(filePath);
-        styleRepository.deleteById(id);
-
     }
 }
